Allow interactive selection of parents when adding a subtag
The whole code seems to be designed around the fact that you can select
multiple parents, even though the tag selection is singleSelect. It would
be straightforward to convert this to a ComboBox. When you select multiple
parents, a copy of the tag is added into each parent.

This feature is not really requested, but a) its not completely useless,
b) and the whole edit subtag with tagset selection and parent tag
selection actually need most of that code, might has well try to develop
and test it first in a separate block.